3-7-23 Update
Fixed the issues with the sequential auto not working, but issues with resetting drive and moving on the ramp still persist.

Also added a new GoToPoseRelative method which makes the robot travel to the desired pose based off the position in which it's called, rather than the absolute position of the field.

Added a new parameter to the GoToPose and GoToPoseRelative methods which takes the desired drive power for the sequence.